requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Authlib==1.4.1
WTForms==3.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
